requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
opencv-python-headless==4.7.0.72
pillow>=10.1
scikit-image==0.25.0
python-multipart==0.0.6
pytest>=8.2,<9
pytest-asyncio>=0.26.0
httpx==0.24.0
pytesseract==0.3.13
numpy<2.0.0